ntheory: rename residue.py to residue_ntheory.py
So that it doesn't collide with the residue() function from complex residues.
