+++ sympy/ntheory/__init__.py
@@ -8,7 +8,7 @@ from primetest import isprime, mr
 from factor_ import divisors, factorint, multiplicity, perfect_power,\
 pollard_pm1, pollard_rho, primefactors, totient, trailing, divisor_count
 from partitions_ import npartitions
-from residue import is_primitive_root, is_quad_residue, legendre_symbol,\
-n_order, totient_
+from residue_ntheory import is_primitive_root, is_quad_residue, \
+    legendre_symbol, n_order, totient_
 from multinomial import binomial_coefficients, binomial_coefficients_list,\
 multinomial_coefficients
