+++ sympy/ntheory/residue_ntheory.py
